# Assignment 2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GitHub Issues Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/webhook": {
      "post": {
        "summary": "Github Webhook",
        "description": "Receive and validate GitHub webhook deliveries.",
        "operationId": "receive_github_webhook",
        "responses": {
          "204": {
            "description": "Webhook delivery accepted and processed successfully"
          },
          "400": {
            "description": "Invalid or unsupported webhook delivery",
            "content": {
              "application/json": {
                "examples": {
                  "unsupported_event": {
                    "summary": "Unsupported GitHub event",
                    "value": {
                      "error": "unsupported_event",
                      "message": "Unsupported GitHub webhook event.",
                      "status_code": 400
                    }
                  },
                  "invalid_payload": {
                    "summary": "Invalid JSON payload",
                    "value": {
                      "error": "invalid_payload",
                      "message": "Webhook payload is not valid JSON.",
                      "status_code": 400
                    }
                  },
                  "unsupported_action": {
                    "summary": "Unsupported webhook action",
                    "value": {
                      "error": "unsupported_action",
                      "message": "Unsupported action 'unknown' for event 'issues'.",
                      "status_code": 400
                    }
                  },
                  "missing_delivery_id": {
                    "summary": "Missing delivery ID",
                    "value": {
                      "error": "missing_delivery_id",
                      "message": "X-GitHub-Delivery header is required.",
                      "status_code": 400
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Webhook signature validation failed",
            "content": {
              "application/json": {
                "example": {
                  "error": "invalid_signature",
                  "message": "Webhook signature validation failed.",
                  "status_code": 401
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "required": true,
            "description": "GitHub HMAC SHA-256 signature used to verify the webhook payload.",
            "schema": {
              "type": "string",
              "example": "sha256=<signature>"
            }
          },
          {
            "name": "X-GitHub-Event",
            "in": "header",
            "required": true,
            "description": "GitHub webhook event type, such as issues, issue_comment, or ping.",
            "schema": {
              "type": "string",
              "example": "issues"
            }
          },
          {
            "name": "X-GitHub-Delivery",
            "in": "header",
            "required": false,
            "description": "Unique GitHub delivery identifier. Required for issues and issue_comment events and used for idempotent event storage.",
            "schema": {
              "type": "string",
              "example": "12345678-abcd-1234-abcd-123456789abc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "action": "opened",
                "issue": {
                  "number": 4
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response returned by the API.",
        "examples": [
          {
            "error": "github_api_error",
            "message": "Issue not found",
            "status_code": 404
          }
        ]
      },
...
    }
  }
}